made private node

+++ server/src/controllers/notes.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from "express";
 import NoteModel from '../models/note';
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUser = req.session.userId;
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUser)
+        const notes = await NoteModel.find({ userId: authenticatedUser }).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error)
@@ -14,13 +17,18 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.id;
+    const authenticatedUser = req.session.userId;
     try {
+        assertIsDefined(authenticatedUser)
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(404, 'Invalid note id!')
         }
         const note = await NoteModel.findById(noteId).exec();
         if (!note) {
             throw createHttpError(404, 'Note not found!')
+        }
+        if (!note.userId.equals(authenticatedUser)) {
+            throw createHttpError(401, 'You cannot access this note!')
         }
         res.status(200).json(note);
     } catch (error) {
@@ -29,6 +37,7 @@ export const getNote: RequestHandler = async (req, res, next) => {
 }
 
 interface CreateNoteBody {
+    // userId?: 
     title?: string,
     text?: string
 }
@@ -36,12 +45,14 @@ interface CreateNoteBody {
 //we put unkown because these are other parameneter that are not needed, it restrics it
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const { title, text } = req.body;
+    const authenticatedUser = req.session.userId;
     try {
+        assertIsDefined(authenticatedUser)
         if (!title) {
             throw createHttpError(400, 'Note must have a title!')
         }
 
-        const newNote = await NoteModel.create({ title, text });
+        const newNote = await NoteModel.create({ title, text, userId: authenticatedUser });
         res.status(201).json(newNote);
     } catch (error) {
         next(error)
@@ -59,7 +70,9 @@ interface UpdateNoteBody {
 export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBody, unknown> = async (req, res, next) => {
     const noteId = req.params.id;
     const { title, text } = req.body;
+    const authenticatedUser = req.session.userId;
     try {
+        assertIsDefined(authenticatedUser)
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(404, 'Invalid note id!')
         }
@@ -73,6 +86,9 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
         if (!note) {
             throw createHttpError(404, 'Note not found!')
         }
+        if (!note.userId.equals(authenticatedUser)) {
+            throw createHttpError(401, 'You cannot access this note!')
+        }
         note.title = title;
         note.text = text;
         const updatedNote = await note.save();
@@ -84,7 +100,9 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.id;
+    const authenticatedUser = req.session.userId;
     try {
+        assertIsDefined(authenticatedUser)
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(404, 'Invalid note id!')
         }
@@ -94,7 +112,9 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
         if (!note) {
             throw createHttpError(404, 'Note not found!')
         }
-
+        if (!note.userId.equals(authenticatedUser)) {
+            throw createHttpError(401, 'You cannot access this note!')
+        }
         await note.deleteOne();
         res.sendStatus(204);
     } catch (error) {
